+++ be/src/main/java/com/niels/homebanking/entity/Currency.java
@@ -4,6 +4,7 @@ import com.niels.homebanking.entity.extra.CommonEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -20,20 +21,20 @@ public class Currency extends CommonEntity {
     private String symbol;
 
     @Column(nullable = false)
-    private long exchange;
+    private BigDecimal exchange;
 
     //
     public Currency() {
     }
 
-    public Currency(String name, String iso, String symbol, long exchange) {
+    public Currency(String name, String iso, String symbol, BigDecimal exchange) {
         this.name = name;
         this.iso = iso;
         this.symbol = symbol;
         this.exchange = exchange;
     }
 
-    public Currency(UUID id, LocalDateTime createdAt, LocalDateTime updatedAt, String name, String iso, String symbol, long exchange) {
+    public Currency(UUID id, LocalDateTime createdAt, LocalDateTime updatedAt, String name, String iso, String symbol, BigDecimal exchange) {
         super(id, createdAt, updatedAt);
         this.name = name;
         this.iso = iso;
@@ -66,11 +67,11 @@ public class Currency extends CommonEntity {
         this.symbol = symbol;
     }
 
-    public long getExchange() {
+    public BigDecimal getExchange() {
         return exchange;
     }
 
-    public void setExchange(long exchange) {
+    public void setExchange(BigDecimal exchange) {
         this.exchange = exchange;
     }
